Fixing create release script

+++ create-release-zip.sh
@@ -36,7 +36,10 @@ VERSION="$(cat version.txt)-$(cat buildnumber.txt)"
 #echo "$VERSION" > $TMP_RELEASE_FOLDER/full-version.txt
 echo "$VERSION" > full-version.txt
 
-rm $RELEASES_FOLDER -R
+if [ -f "releases" ]; then
+  echo"  Removing releases folder..."
+  rm $RELEASES_FOLDER -R
+fi
 mkdir -p $RELEASES_FOLDER
 
 #cd .tmp/BridgeArduinoSerialToMqttSplitCsv
